[flang][OpenMP][OpenACC] remove libEvaluate dependency in passes
